requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
requests
PyPDF2
python-multipart
python-dotenv
